RatingListView moved to 'profile' package
