+++ src/main/java/egl/client/controller/profile/global/GlobalProfilesController.java
@@ -9,7 +9,7 @@ import java.util.stream.Collectors;
 
 import egl.client.controller.profile.CredentialsInfoController;
 import egl.client.controller.profile.ProfileSelectController;
-import egl.client.controller.profile.info.RatingListView;
+import egl.client.controller.profile.RatingListView;
 import egl.client.model.core.profile.Credentials;
 import egl.client.model.core.profile.Profile;
 import egl.client.model.core.statistic.ProfileStatistic;
